Bug 812956 - Implement FFOS Privacy Policy r=pike r=arcturus r=kaze
This patch implements a URL loader in FTU, with an iFrame, because we don't want
that the user could go outside of the FTU yet.

If there is no connection, we display an error dialog instead, with the link
written so that the user could type it on another computer.

I add to change how the panels are made fullscreen because of the iFrame, which
needs its parent nodes' dimensions to be defined without percentages.

We fix a small bug in the l10n library as well: it didn't replace the keys with
a dash or a colon in them.