Fixes an issue where the unread marker was not added properly in some situations
STR was: be in a thread, have the SMS app in background, receive a message on
this thread, receive a message on another thread, tap the notification for this
another thread, then go back: the unread marker was not appearing for the
previous thread. But restarting the app was making it appear.